Move OldHashIndices/OldHashDictionaries to contrib/
