+++ benchmark/bench_indices.jl
@@ -3,11 +3,12 @@ module BenchHashIndices
 using BenchmarkTools
 using Dictionaries
 using OrderedCollections
+# include("../contrib/OldHashDictionaries.jl")
+# using .OldHashDictionaries
 
 const suite = BenchmarkGroup()
 
-#sizes = [(8 .^ (0:8))...]
-sizes = [10, 100, 1000, 10_000] #, 10_000, 10_000_000]
+sizes = [10, 100, 1000, 10_000]
 cutoff = 101
 
 function build_vector_by_insertion(n)
@@ -50,13 +51,13 @@ function build_hashindices_by_insertion(n)
     return out
 end
 
-function build_old_hashindices_by_insertion(n)
-    out = Dictionaries.OldHashIndices{Int}()
-    for i in 1:n
-        insert!(out, i)
-    end
-    return out
-end
+# function build_old_hashindices_by_insertion(n)
+#     out = OldHashIndices{Int}()
+#     for i in 1:n
+#         insert!(out, i)
+#     end
+#     return out
+# end
 
 function empty_by_deletion(set::Vector, n)
     for i in 1:n
@@ -300,53 +301,53 @@ function basic_hash_indices_test(N)
     return out
 end
 
-function basic_old_hash_indices_test(N)
-    h = Dictionaries.OldHashIndices{Int}()
-    out = true
-    for i in 1:N
-        insert!(h, i)
-    end
-    for i in 1:N
-        out &= i in h
-    end
-    for i in 1:2:N
-        delete!(h, i)
-    end
-    for i in 1:N
-        out &= (i in h) == iseven(i)
-    end
-    for i in 1:2:N
-        insert!(h, i)
-    end
-    for i in 1:N
-        out &= i in h
-    end
-    for i in 1:N
-        delete!(h, i)
-    end
-    out &= isempty(h)
-    insert!(h, 7)
-    out &= 7 in h
-    for i in 1:N
-        set!(h, i)
-    end
-    for i in 1:N
-        out &= i in h
-    end
-    for i in 1:2:N
-        delete!(h, i)
-    end
-    for i in 1:N
-        out &= (i in h) == iseven(i)
-    end
-    for i in N+1:2N
-        insert!(h, i)
-    end
-    for i in 1:2N
-        out &= (i in h) == (i > N || iseven(i))
-    end
-    return out
-end
+# function basic_old_hash_indices_test(N)
+#     h = OldHashIndices{Int}()
+#     out = true
+#     for i in 1:N
+#         insert!(h, i)
+#     end
+#     for i in 1:N
+#         out &= i in h
+#     end
+#     for i in 1:2:N
+#         delete!(h, i)
+#     end
+#     for i in 1:N
+#         out &= (i in h) == iseven(i)
+#     end
+#     for i in 1:2:N
+#         insert!(h, i)
+#     end
+#     for i in 1:N
+#         out &= i in h
+#     end
+#     for i in 1:N
+#         delete!(h, i)
+#     end
+#     out &= isempty(h)
+#     insert!(h, 7)
+#     out &= 7 in h
+#     for i in 1:N
+#         set!(h, i)
+#     end
+#     for i in 1:N
+#         out &= i in h
+#     end
+#     for i in 1:2:N
+#         delete!(h, i)
+#     end
+#     for i in 1:N
+#         out &= (i in h) == iseven(i)
+#     end
+#     for i in N+1:2N
+#         insert!(h, i)
+#     end
+#     for i in 1:2N
+#         out &= (i in h) == (i > N || iseven(i))
+#     end
+#     return out
+# end
 
 for n in sizes
     r = 1:n
@@ -380,9 +381,9 @@ for n in sizes
     even_hash_indices = HashIndices(2:2:n)
     odd_hash_indices = HashIndices(1:2:n)
 
-    old_hash_indices = Dictionaries.OldHashIndices(r)
-    even_old_hash_indices = Dictionaries.OldHashIndices(2:2:n)
-    odd_old_hash_indices = Dictionaries.OldHashIndices(1:2:n)
+    # old_hash_indices = OldHashIndices(r)
+    # even_old_hash_indices = OldHashIndices(2:2:n)
+    # odd_old_hash_indices = OldHashIndices(1:2:n)
 
     suite_n = suite["$n"] = BenchmarkGroup()
 
@@ -392,7 +393,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable OrderedSet($r)
     n < cutoff && (s["Indices"] = @benchmarkable Indices($r))
     s["HashIndices"] = @benchmarkable HashIndices($r)
-    s["OldHashIndices"] = @benchmarkable HashIndices($r)
+    #s["OldHashIndices"] = @benchmarkable HashIndices($r)
 
     s = suite_n["build by insertion ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector (push!)"] = @benchmarkable build_set_by_insertion($n))
@@ -400,7 +401,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable build_ordered_set_by_insertion($n)
     n < cutoff && (s["Set"] = @benchmarkable build_indices_by_insertion($n))
     s["HashIndices"] = @benchmarkable build_hashindices_by_insertion($n)
-    s["OldHashIndices"] = @benchmarkable build_old_hashindices_by_insertion($n)
+    #s["OldHashIndices"] = @benchmarkable build_old_hashindices_by_insertion($n)
 
     s = suite_n["copy ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector (pop!)"] = @benchmarkable copy($vec))
@@ -408,7 +409,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable copy($ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable copy($indices))
     s["HashIndices"] = @benchmarkable copy($hash_indices)
-    s["OldHashIndices"] = @benchmarkable copy($old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable copy($old_hash_indices)
 
     s = suite_n["copy and empty by deletion ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector (pop!)"] = @benchmarkable empty_by_deletion(copy($vec), $n))
@@ -416,21 +417,21 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable empty_by_deletion(copy($ordered_set), $n)
     n < cutoff && (s["Indices"] = @benchmarkable empty_by_deletion(copy($indices), $n))
     s["HashIndices"] = @benchmarkable empty_by_deletion(copy($hash_indices), $n)
-    s["OldHashIndices"] = @benchmarkable empty_by_deletion(copy($old_hash_indices), $n)
+    #s["OldHashIndices"] = @benchmarkable empty_by_deletion(copy($old_hash_indices), $n)
 
     s = suite_n["insertion/deletion tests ($n)"] = BenchmarkGroup()
     s["Set"] = @benchmarkable basic_set_test($n)
     s["OrderedSet"] = @benchmarkable basic_ordered_set_test($n)
     n < cutoff && (s["Indices"] = @benchmarkable basic_indices_test($n))
     s["HashIndices"] = @benchmarkable basic_hash_indices_test($n)
-    s["OldHashIndices"] = @benchmarkable basic_old_hash_indices_test($n)
+    #s["OldHashIndices"] = @benchmarkable basic_old_hash_indices_test($n)
 
     s = suite_n["insertion/deletion tests ($n)"] = BenchmarkGroup()
     s["Set"] = @benchmarkable basic_set_test($n)
     s["OrderedSet"] = @benchmarkable basic_ordered_set_test($n)
     n < cutoff && (s["Indices"] = @benchmarkable basic_indices_test($n))
     s["HashIndices"] = @benchmarkable basic_hash_indices_test($n)
-    s["OldHashIndices"] = @benchmarkable basic_old_hash_indices_test($n)
+    #s["OldHashIndices"] = @benchmarkable basic_old_hash_indices_test($n)
 
     s = suite_n["unique/distinct (high uniqueness, unsorted) ($n)"] = BenchmarkGroup()
     s["Vector (unique)"] = @benchmarkable unique($mostly_unique)
@@ -462,7 +463,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable all_in($ordered_set, $n)
     n < cutoff && (s["Indices"] = @benchmarkable all_in($indices, $n))
     s["HashIndices"] = @benchmarkable all_in($hash_indices, $n)
-    s["OldHashIndices"] = @benchmarkable all_in($old_hash_indices, $n)
+    #s["OldHashIndices"] = @benchmarkable all_in($old_hash_indices, $n)
 
     s = suite_n["not in ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable not_in($vec, $n))
@@ -470,7 +471,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable not_in($ordered_set, $n)
     n < cutoff && (s["Indices"] = @benchmarkable not_in($indices, $n))
     s["HashIndices"] = @benchmarkable not_in($hash_indices, $n)
-    s["OldHashIndices"] = @benchmarkable not_in($old_hash_indices, $n)
+    #s["OldHashIndices"] = @benchmarkable not_in($old_hash_indices, $n)
 
     s = suite_n["count ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable count(iseven, $vec))
@@ -478,7 +479,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable count(iseven, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable count(iseven, $indices))
     s["HashIndices"] = @benchmarkable count(iseven, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable count(iseven, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable count(iseven, $old_hash_indices)
 
     s = suite_n["sum ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable sum($vec))
@@ -486,7 +487,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable sum($ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable sum($indices))
     s["HashIndices"] = @benchmarkable sum($hash_indices)
-    s["OldHashIndices"] = @benchmarkable sum($old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable sum($old_hash_indices)
 
     s = suite_n["foreach ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable foreachsum($vec))
@@ -494,7 +495,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable foreachsum($ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable foreachsum($indices))
     s["HashIndices"] = @benchmarkable foreachsum($hash_indices)
-    s["OldHashIndices"] = @benchmarkable foreachsum($old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable foreachsum($old_hash_indices)
 
     s = suite_n["filter-map-reduce via generator ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable sum($(2x for x in vec if isodd(x))))
@@ -502,7 +503,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable sum($(2x for x in ordered_set if isodd(x)))
     n < cutoff && (s["Indices"] = @benchmarkable sum($(2x for x in indices if isodd(x))))
     s["HashIndices"] = @benchmarkable sum($(2x for x in hash_indices if isodd(x)))
-    s["OldHashIndices"] = @benchmarkable sum($(2x for x in old_hash_indices if isodd(x)))
+    #s["OldHashIndices"] = @benchmarkable sum($(2x for x in old_hash_indices if isodd(x)))
 
     s = suite_n["filter (most) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter($pred1, $vec))
@@ -510,7 +511,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter($pred1, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable filter($pred1, $indices))
     s["HashIndices"] = @benchmarkable filter($pred1, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable filter($pred1, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable filter($pred1, $old_hash_indices)
 
     s = suite_n["filter (half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter(iseven, $vec))
@@ -518,7 +519,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter(iseven, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable filter(iseven, $indices))
     s["HashIndices"] = @benchmarkable filter(iseven, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable filter(iseven, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable filter(iseven, $old_hash_indices)
 
     s = suite_n["filter (few) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter($pred2, $vec))
@@ -526,7 +527,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter($pred2, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable filter($pred2, $indices))
     s["HashIndices"] = @benchmarkable filter($pred2, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable filter($pred2, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable filter($pred2, $old_hash_indices)
 
     s = suite_n["copy and filter! (most) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter!($pred1, copy($vec)))
@@ -534,7 +535,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter!($pred1, copy($ordered_set))
     n < cutoff && (s["Indices"] = @benchmarkable filter!($pred1, copy($indices)))
     s["HashIndices"] = @benchmarkable filter!($pred1, copy($hash_indices))
-    s["OldHashIndices"] = @benchmarkable filter!($pred1, copy($old_hash_indices))
+    #s["OldHashIndices"] = @benchmarkable filter!($pred1, copy($old_hash_indices))
 
     s = suite_n["copy and filter! (half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter!(iseven, copy($vec)))
@@ -542,7 +543,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter!(iseven, copy($ordered_set))
     n < cutoff && (s["Indices"] = @benchmarkable filter!(iseven, copy($indices)))
     s["HashIndices"] = @benchmarkable filter!(iseven, copy($hash_indices))
-    s["OldHashIndices"] = @benchmarkable filter!(iseven, copy($old_hash_indices))
+    #s["OldHashIndices"] = @benchmarkable filter!(iseven, copy($old_hash_indices))
 
     s = suite_n["copy and filter! (few) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable filter!($pred2, copy($vec)))
@@ -550,7 +551,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable filter!($pred2, copy($ordered_set))
     n < cutoff && (s["Indices"] = @benchmarkable filter!($pred2, copy($indices)))
     s["HashIndices"] = @benchmarkable filter!($pred2, copy($hash_indices))
-    s["OldHashIndices"] = @benchmarkable filter!($pred2, copy($old_hash_indices))
+    #s["OldHashIndices"] = @benchmarkable filter!($pred2, copy($old_hash_indices))
 
     s = suite_n["union ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable union($even_vec, $odd_vec))
@@ -558,7 +559,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable union($even_ordered_set, $odd_ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable union($even_indices, $even_indices))
     s["HashIndices"] = @benchmarkable union($even_hash_indices, $odd_hash_indices)
-    s["OldHashIndices"] = @benchmarkable union($even_old_hash_indices, $odd_old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable union($even_old_hash_indices, $odd_old_hash_indices)
 
     s = suite_n["intersect (empty) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable intersect($even_vec, $odd_vec))
@@ -566,7 +567,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable intersect($even_ordered_set, $odd_ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable intersect($even_indices, $odd_indices))
     s["HashIndices"] = @benchmarkable intersect($even_hash_indices, $odd_hash_indices)
-    s["OldHashIndices"] = @benchmarkable intersect($even_old_hash_indices, $odd_old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable intersect($even_old_hash_indices, $odd_old_hash_indices)
 
     s = suite_n["intersect (half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable intersect($even_vec, $vec))
@@ -574,7 +575,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable intersect($even_ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable intersect($even_indices, $indices))
     s["HashIndices"] = @benchmarkable intersect($even_hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable intersect($even_old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable intersect($even_old_hash_indices, $old_hash_indices)
 
     s = suite_n["intersect (whole) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable intersect($vec, $vec))
@@ -582,7 +583,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable intersect($ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable intersect($indices, $indices))
     s["HashIndices"] = @benchmarkable intersect($hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable intersect($old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable intersect($old_hash_indices, $old_hash_indices)
 
     s = suite_n["setdiff (whole) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable setdiff($even_vec, $odd_vec))
@@ -590,7 +591,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable setdiff($even_ordered_set, $odd_ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable setdiff($even_indices, $odd_indices))
     s["HashIndices"] = @benchmarkable setdiff($even_hash_indices, $odd_hash_indices)
-    s["OldHashIndices"] = @benchmarkable setdiff($even_old_hash_indices, $odd_old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable setdiff($even_old_hash_indices, $odd_old_hash_indices)
 
     s = suite_n["setdiff (half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable setdiff($even_vec, $vec))
@@ -598,7 +599,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable setdiff($even_ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable setdiff($even_indices, $indices))
     s["HashIndices"] = @benchmarkable setdiff($even_hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable setdiff($even_old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable setdiff($even_old_hash_indices, $old_hash_indices)
 
     s = suite_n["setdiff (empty) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable setdiff($vec, $vec))
@@ -606,7 +607,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable setdiff($ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable setdiff($indices, $indices))
     s["HashIndices"] = @benchmarkable setdiff($hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable setdiff($old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable setdiff($old_hash_indices, $old_hash_indices)
 
     s = suite_n["symdiff (whole) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable symdiff($even_vec, $odd_vec))
@@ -614,7 +615,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable symdiff($even_ordered_set, $odd_ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable symdiff($even_indices, $odd_indices))
     s["HashIndices"] = @benchmarkable symdiff($even_hash_indices, $odd_hash_indices)
-    s["OldHashIndices"] = @benchmarkable symdiff($even_old_hash_indices, $odd_old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable symdiff($even_old_hash_indices, $odd_old_hash_indices)
 
     s = suite_n["symdiff (left half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable symdiff($vec, $odd_vec))
@@ -622,7 +623,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable symdiff($ordered_set, $odd_ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable symdiff($indices, $odd_indices))
     s["HashIndices"] = @benchmarkable symdiff($hash_indices, $odd_hash_indices)
-    s["OldHashIndices"] = @benchmarkable symdiff($old_hash_indices, $odd_old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable symdiff($old_hash_indices, $odd_old_hash_indices)
 
     s = suite_n["symdiff (right half) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable symdiff($even_vec, $vec))
@@ -630,7 +631,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable symdiff($even_ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable symdiff($even_indices, $indices))
     s["HashIndices"] = @benchmarkable symdiff($even_hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable symdiff($even_old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable symdiff($even_old_hash_indices, $old_hash_indices)
 
     s = suite_n["symdiff (empty) ($n)"] = BenchmarkGroup()
     n < cutoff && (s["Vector"] = @benchmarkable symdiff($vec, $vec))
@@ -638,7 +639,7 @@ for n in sizes
     s["OrderedSet"] = @benchmarkable symdiff($ordered_set, $ordered_set)
     n < cutoff && (s["Indices"] = @benchmarkable symdiff($indices, $indices))
     s["HashIndices"] = @benchmarkable symdiff($hash_indices, $hash_indices)
-    s["OldHashIndices"] = @benchmarkable symdiff($old_hash_indices, $old_hash_indices)
+    #s["OldHashIndices"] = @benchmarkable symdiff($old_hash_indices, $old_hash_indices)
 end
 
 end  # module
